Intergrated pdf automatic request and get
